Allow tests to be skipped if full regex support is missing
Summary: GCC 4.8 did not support full regex but did include the functions.  If we try to run some of our tests that intentionally cause failures the system crashes when built with GCC 4.8.  Add the ability to detect what compiler and version the system was built with so we can skip tests that are known to fail without full regex support

Test Plan: Run rocksdb.collation on the regular build and see that it passes.  Then run it on the --gcc48 build and see that it gets skipped.

Reviewers: hermanlee4, mung, gunnarku

Reviewed By: mung

Subscribers: webscalesql-eng

Closes https://github.com/facebook/mysql-5.6/pull/399
fbshipit-source-id: 7b70e6d

Differential Revision: https://reviews.facebook.net/D64515
Differential Revision: D4134960