Fix test failures
Summary:
During recovery thd is null, so we need to guard against this. It is also necessary
to check the query that is getting reset is the same one that is set while applying
the rows query log event.

Squash with: 42bdfc801d579722efc5d2d9eedc9c37cd7c9fe0

Test Plan: mtr