Fix test failures
Summary:
During recovery thd is null, so we need to guard against this. I think it is also
better to check for the current we are resetting thd query value which was set
when applying the same rows query log event.

Squash with: 42bdfc801d579722efc5d2d9eedc9c37cd7c9fe0
Closes https://github.com/facebook/mysql-5.6/pull/367

Reviewed By: lth

Differential Revision: D4099803

Pulled By: santoshbanda

fbshipit-source-id: 4b4a041