RBR's idempotent mode to make slave crash safe
Summary:
Currently, to ensure crash safety on slaves a DB->GTID map is maintained
which is updated after every commit. The map is required to identify holes
in the sequence of gtids caused by using MTS. In RBR we can make use of
idempotency of row events to handle holes instead. We only track the max gtid
encountered and after a crash we turn on idempotency for all transactions which
have gtid <= max gtid. If any trx was already executed in this range it's okay
because we've turned on idempotent mode.

Test Plan:
Created a new suite rpl_recovery with some existing and some new test cases.
Added a new combination in this suite to run in idempotent recovery mode.