Fix deadlock detection tests (#393) (#393)
Summary:
Checking "Waiting for row locks" is not reliable because that message will display even if the thread is not blocked. One way to work around this is to use the newly created rocksdb_locks table to see if all the locks we expect the transaction to take has been taken. The better way is to use rocksdb_trx once we have that.
Closes https://github.com/facebook/mysql-5.6/pull/393

Differential Revision: D4128101 (https://github.com/facebook/mysql-5.6/commit/9addd7ac2d1ca30a5c85c42250b8c797d22f9558)

Pulled By: lth

fbshipit-source-id: 6eeae62ef51