Fix rocksdb.collation test
Summary: The test used grep to look for an expected line in the log file.  Somehow (maybe because of debugging) the "-A 1" was included that told grep to output the found line plus one more line.  Normally the found line would be the last line in the file, but if an index was being dropped, sometimes the "index dropped" message would get in after the "Invalid pattern" message and would get outputted as well.  This will fix the arguments to grep so that the extra line will no longer be printed.

Squash with: https://reviews.facebook.net/D64515

Test Plan: MTR