[sync client ssl] Ensure socket is in NON BLOCKING mode when SSL is …
Summary:
…used

If the socket is in BLOCKING mode when ssl connect is performed, it can
block indifenitely if the server does not respond.

Force NON BLOCKING mode before sslconnect() is called.

Put back setting the socket to NON BLOCKING in vio_ssl_set_blocking().
It was changed in https://reviews.facebook.net/D51801. The reason stated
there is not very clear, but tests didn't show this is a problem.
Closes https://github.com/facebook/mysql-5.6/pull/404

Reviewed By: abal147

Differential Revision: D4152341

Pulled By: anca-agape

fbshipit-source-id: 2438810