Heartbeat event should carry the latest master timestamp
Summary: Heartbeat events should carry the now() timestamp from the master and
the last_master_timstamp from the slave. In other words, HB should always carry
the lastet master timestamp. This will help services connected to slaves know
if they are lagging even when no real events are flowing in the system. Also,
for this to work last_master_timestamp should be updated whenever a HB event is
received.

Test Plan: Added a new test to check this behavior.