[mysql-cli] Fix handle_kill_signal to use SSL
Summary: Currently the kill signal handler does not use SSL, even when
the correct options are passed. This causes it to fail to kill a running
query that was started on an SSL connection because it does not have a
password to connect with and does not perform an SSL connection with the
certificate.

Test Plan: Run mysql CLI with an SSL connection, execute SELECT
SLEEP(500); then Ctrl-C the process

Reviewers: abalsara, mung

Subscribers:

Tasks: 14516640

Blame Revision: