Add linter to validate .test/.result files
Summary: We need a new linter to validate that if a new .test file is added the matching .result file is also added.  This new linter will check any time a .test or .result file is included in a diff to validate that the opposite file exists.

Test Plan: N/A