RBR's idempotent mode to make slave crash safe (#384)
Summary:
Currently to ensure crash safety in slave a DB->GTID map is maintained
which is updated after every commit. The map is required to identify holes
in the sequence of gtids caused by using MTS. In RBR we can make use of
idempotency of row events to handle holes instead. We only track the max gtid
encountered and after a crash we turn on idempotency for all transactions which
have gtid less than the max gtid. If any trx was already executed in this range
it's okay because we've turned on idempotent mode.
Closes https://github.com/facebook/mysql-5.6/pull/384

Differential Revision: D4113718

Pulled By:
abhinav04sharma

fbshipit-source-id: 4c6da14

fbshipit-source-id: 253c185fc6c