Yield the background thread when not necessary (#465)
Summary:
These changes are the proposed fix for
https://github.com/facebook/mysql-5.6/issues/465. The intent of the
background thread is to wake up once per second and call SyncWAL() if
all the preconditions are met.

Previous implementation ended up constantly looping because of a logic
error. This solution removes the unnecessary checks and uses a call to
`mysql_cond_timedwait()` (ends up being `pthread_cond_timedwait()`) to
make sure that thread yields its quantum and OS scheduler can do its
thing while we're waiting for the timer to expire.

We still loop, but for the majority of the time we end up sleeping and
not using the CPU.

Basic testing:

  - Verify that function body is called once per second. If we need then
    we can tune it to be more precise and use what approximate amount of
    ms-s or ns-s we desire.
  - Before the fix the code path was executed hundreds of time per
    second and 25-49% of CPU was utilized. After the fix `htop` show
    marginal (0.0%) utilization.
  - `perf
Closes https://github.com/facebook/mysql-5.6/pull/467

Differential Revision: D4294886

Pulled By: gunnarku

fbshipit-source-id: 0bf7d4c