Fix errors in TC_LOG_MMAP
Summary:
MySQL 5.6 has some errors in the implementation of TC_LOG_MMAP that were resolved in 5.7.  These errors show up when attempting to alter a table to change its storage engine from InnoDB to MyRocks or vice versa.  See also: https://www.percona.com/blog/2013/10/23/problems-with-multiple-xa-storage-engines-in-mysql-5-6/

This diff (at least the code change pieces) should be removed if/when MyRocks is ported to 5.7 or later.

This diff also contains the beginnings of a new suite of tests that intentionally run with two transactional engines at the same time (InnoDB an MyRocks).
Closes https://github.com/facebook/mysql-5.6/pull/454

Reviewed By: lth

Differential Revision: D4238463

Pulled By: jkedgar

fbshipit-source-id: 7517b13