Improved diagnostics to understand MyRocks execution
Interface MySQL provides for implementing custom storage engines is
quite cumbersome. For the purposes of the debugging and understanding
the flow of execution we need to be able to clearly observe and trace
how MyRocks code is being called. Interface itself is described in
`sql\handler.h`.

As a standard solution, MySQL exposes DBUG package and various macros
associated with it.

This set of changes makes sure that every call to any of the functions
MyRocks implements will be tracked with a pair of `DBUG_ENTER_FUNC` and
`DBUG_RETURN*` macros. We can now change DBUG settings during runtime
(`SET GLOBAL debug = '+d,enter,exit:L';` or something similar) and
observe the execution flow.

`DBUG_ENTER_FUNC` is a simple customization based on `DBUG_ENTER` and it
uses `__PRETTY_FUNCTION__` to delegate the work of composing correct
function names to a compiler. Yes, this functionality is specific to
`Clang` or `GCC` and these are the only toolsets we care about now.

For future references: please use `DBUG_ENTER_FUNC` always at the top of
the function as opposed to after variable declarations (which may cause
tons of code in some c'tor to be executed) to make sure that logs will
correctly reflect the reality.

As a bonus we also add some extra invariants and `override` specifier to
all the applicable methods :-)