avoid calling GetOptions() destructor per query (#503) (#503)
Summary:
We currently call GetOptions() for every query by MyRocks, to determine if we can use a bloom filter. GetOptions() constructs an object, which we can avoid doing if we cache the prefix_extractor (which is immutable for the lifetime of the db) when setting up key descriptors.
Closes https://github.com/facebook/mysql-5.6/pull/503

Differential Revision: D4397391

Pulled By: alxyang