Fix heuristic for using ref access
Summary: In MySQL, a heuristic exists that enforces using an index ref access being used over full table scan, when the query is considered covering. The check for covering keys does not work for clustered primary keys though, because MySQL assumes that accessing any column not in the key definition will be inefficient.

Reviewed By: tianx

Differential Revision: D4561446

fbshipit-source-id: e4a9d63