Validate DATA and INDEX DIRECTORY usage for CREATE TABLE in MyRocks
Summary:
This fixes #532. `DATA DIRECTORY` and `INDEX DIRECTORY` areu sed to
create tables under a specific location outside the MySQL data
directory. We don't support this for MyRocks.

The `rocksdb_datadir` setting should be used to configure RocksDB data
directory.

As an extra bonus we also run `clang-format` on the code and fix the
minor issues introduced recently.

Differential Revision: D4574100 (https://github.com/facebook/mysql-5.6/commit/71fba268a7b6ea342fd63768ed314493da248db0)

fbshipit-source-id: 79d433cf4ce