Generate unique filenames for rocksdb_bulk_load
Summary:
When rocksdb_bulk_load is enabled, the filename generated is based only
on the table and index. However, if two client threads try to load the
same table concurrently, the names collide and one overwrites the other.
Add an atomic counter that is incremented for each file generated to
generate unique filename.

Differential Revision: D4586801 (https://github.com/facebook/mysql-5.6/commit/7dd63734529d11b728f8d09d54f0fa92ad3b4673)

fbshipit-source-id: b898a8b656a