Don't assume buffer is relative to table->record[0]
Summary:  Some of the functions that unpack and pack a record assume that the buffer passed in is a part of table->record[0].  This appears to be false when using partitions.  This diff updates this code to not make that assumption.  To do this, the field->ptr value is temporarily updated to point into the provided buffer and then restored.  In order to avoid potential problems with this (i.e. field->ptr left pointing into the buffer accidently), code was refactored to keep the location where field->ptr was reset and where it was restored close together.

Test Plan: MTR plus new test for Issue 495