Add zstd compression option for replication
Summary:
This diff adds a generic protocol option to modify the compression
algorithm at connect time. This is done using the 'compression_lib' connection
attribute, which when combined with the compress flag, will determine a new
compression library for network compression. It is meant to be extendable to
other compression libraries.

Protocol info:
When a client wants to use compression, they can either set the compress flag
as they do now, or use mysql_options to select a compression library from the
given enums. Under the hood, this will set the compress flag and pass the
library through to the server via a connection attribute

Rollout:
To change the compression on slave threads connect to the slave and run
'SET slave_compression_lib="zstd";stop slave;start slave;'
This assumes the master also supports the same compression algorithms.
In the event the master is not upgraded an error will occur on the slave side

Reviewed By: lth

Differential Revision: D4511885

fbshipit-source-id: 24a63df