Fix the varbinary encoding to avoid an inefficiency
Summary:
The existing varbinary format has an inefficiency such that any source that has a length of a multiple of 8 bytes will generate an encoded value that has an extra 9 bytes that can be avoided.

The pack/unpack/skip functions needed access to the key's version information, so instead of passing it in as another parameter (which gets ugly), I made all of these functions member functions of Rdb_key_def.

While investigating I realized that for short input values breaking in multiples of 8 bytes causes quite a bit of overhead.  The new scheme breaks the input into 4 byte segments for the first 32 bytes of input and after that it uses 8 byte segments.  This provides a good compromise between the savings that 8 byte segments gives for long inputs and 4 byte segments give for short inputs.

Test Plan: MTR and manual check of format

Reviewers: alexyang, mung, herman

Reviewed By: herman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4637146

Tasks: 15169935

Signature: t1:4637146:1490897813:40568fa25d846767368739275a8e07dd9c7c3937