Setting some *histogram_step_size* variables to NULL crashed the server
Summary: The histogram_step_size* variables had a CHECK function that was not validating that the passed in string was non-NULL before it was used.  This meant that setting the variable to NULL would crash the server.  This diff adds the check and adds a test for each variable.

Test Plan: MTR