Setting some *histogram_step_size* variables to NULL crashed the server
Summary:
The histogram_step_size* variables had a CHECK function that was not validating that the passed in string was non-NULL before it was used.  This meant that setting the variable to NULL would crash the server.  This diff adds the check and adds a test for each variable.
Closes https://github.com/facebook/mysql-5.6/pull/549

Differential Revision: D4620894

Pulled By: jkedgar

fbshipit-source-id: be1b74a