Delete the INDEX_STATISTICS key when the INDEX_INFO key is deleted
Summary: Any time an index is dropped the data is deleted and then the INDEX_INFO key is removed, but the INDEX_STATISTICS key was left behind.  Add code to remove it.

Test Plan: MTR - I can't think of a way of seeing the INDEX_STATISTICS keys in order to validate that they are getting deleted.