Delete the INDEX_STATISTICS key when the INDEX_INFO key is deleted (#553)
Summary:
Any time an index is dropped the data is deleted and then the INDEX_INFO key is removed, but the INDEX_STATISTICS key was left behind.  Add code to remove it.
Closes https://github.com/facebook/mysql-5.6/pull/553

Test Plan: MTR - I can't think of a way of seeing the INDEX_STATISTICS keys in order to validate that they are getting deleted.

Reviewed By: alxyang

Differential Revision: D4623812

Pulled By: jkedgar

fbshipit-source-id: 8144bbec1c9