Move timeout_message() so it is available to unittests
Summary:
The timeout_message() that was added was put in a location that made it unavailble for the unittests as the source file that contained it was not included.  Move the function into sql/sql_string.h and make it inline.

Squash with: D47613

Test Plan: MTR and manually run merge_large_tests-t and merge_small_tests-t

Reviewers: herman, mung, alexyang

Reviewed By: alexyang

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4632140

Signature: t1:4632140:1488327137:c49218ad07429a4f42ba45caf0a6ae2e4c0af7f6