Replace connect() with bind() in mtr port check
Summary:
There are times when the mtr setup fails because the server could not
bind to the given port number allocated during start. Switch the port
check from a connect() to a bind() call to determine if the port is
available. It is possible connect() may fail to connect to a port that
another process had bound, but not listening on. The bind() check should
be stronger than the connect() check.

Reviewed By: gunnarku

Differential Revision: D4754831

fbshipit-source-id: fc5999e