Replace connect() with bind() in mtr port check
Summary:
There are times when the mtr setup fails because the server could not
bind to the given port number allocated during start. Switch the port
check from a connect() to a bind() call to determine if the port is
available. It is possible connect() may fail to connect to a port that
another process had bound, but not listening on. The bind() check should
be stronger than the connect() check.

Test Plan:
  Run mtr tests

Reviewers: jkedgar, gunnarku

Reviewed By: gunnarku

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4754831

Tasks: 16819569, 16819560, 16819565, 16819559, 16819549, 16819562, 16819563, 16819557, 16819551, 16819544, 16819545, 16819550, 16819553, 16819567, 16819555

Signature: t1:4754831:1490203434:7fd1c49c15855d615c1c45250bddd0b83bf69e10