Fixing innodb_dropper_big
Summary:
Statements in this stress test can timeout while acquiring row
locks. The python script that runs this test does not handle this so the
transactions in which statements timeout keep running till the
connection is terminated. The timeout for connection was set to a very
high value (8 hours), this changes the connection timeout to 2 minutes.
This test also caused sandcastle runs to get stuck in the "running"
state.

Test Plan: Check if this test passes and if sandcastle runs correctly

Reviewers: arahut, tianx, gunnarku

Reviewed By: gunnarku

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4767216

Tasks: 16794287

Signature: t1:4767216:1490372504:60db97deb72e76e09245a040f3db7e4b7162a42d