Fixing innodb_dropper_big
Summary:
Statements in this stress test can timeout while acquiring row
locks. The python script that runs this test does not handle this so the
transactions in which statements timeout keep running till the
connection is terminated. The timeout for connection was set to a very
high value (8 hours), this changes the connection timeout to 2 minutes.
This test also caused sandcastle runs to get stuck in the "running"
state.

Reviewed By: gunnarku

Differential Revision: D4767216

fbshipit-source-id: 7b49203