Fix rocksdb_sys_vars.rocksdb_force_flush_memtable_and_lzero_now_basic
Summary:
When run as part of the RocksDBBig test set, rocksdb_sys_vars.rocksdb_force_flush_memtable_and_lzero_now_basic was failing.  The problem is that if a compaction is already happening on the specified file a kAborted status is returned, but this is interpreted as a failure by the MyRocks code.

Squash with: D4509351

Test Plan: MTR

Reviewers: alexyang, mung, gunnarku

Reviewed By: mung, gunnarku

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4803800

Signature: t1:4803800:1490896933:debf4b69c5ec5dba9285808fb735bc9887dbd6be