Unconditionally send gtid in SHOW BINARY LOGS WITH GTID
Summary:
When the gtid is empty, nothing is sent over the wire to
the client. This was causing memory issues client side because the
client expects a fixed number of columns per row, and nothing was being sent.
Fix is to send an empty string in it's place.

squash with: 25fec83c7117ca7337a51835082dc55d7007215f

Reviewed By: lth

Differential Revision: D4812840

fbshipit-source-id: 103dd4e