Fix rpl.rpl_unsafe_statements valgrind errors
Summary:
In the test rpl.rpl_unsafe_statements, we are checking to see if a table should be written out to the binlog for RBR based on whether the table was opened for triggers or not. This was implemented in https://github.com/facebook/mysql-5.6/commit/cdeb01291 and was done by setting a flag on the TABLE_LIST object.

However, the lifetime of the TABLE_LIST object is tied to the statement. For statements after a LOCK TABLE statement, we were looking at free'd TABLE_LIST objects, causing errors. To fix, we will skip the check when writing out table map events for locked tables. This means that if tables were locked during a row event, we will always write out the associated table map event, which may cause errors in slave sql thread. We don't write row events under lock tables during OSC though, so this is fine.

Squash with: cdeb01291f4d5440f07a33ad5f53245ed302e617

Reviewed By: abhinav04sharma

Differential Revision: D4824636

fbshipit-source-id: 239279e