Another attempt to fix rpl.rpl_unsafe_statements
Summary:
The rpl.rpl_unsafe_statements test was added in 5.6.29 and when we rebased to 5.6.35 it started causing a problem with the diff "optionally disable binlogging while executing triggers".  This is an attempt to modify that diff to work correctly for this case.

Squash with: cdeb01291f4d5440f07a33ad5f53245ed302e617

Reviewed By: lth

Differential Revision: D4836512

fbshipit-source-id: 87e3650