Fix rocksdb.bloomfilter3
Summary:
Due to changes in RocksDB the rocksdb.bloomfilter3 test was showing errors in Valgrind.  The problem was that the Statistics were being accessed after they were freed, but even fixing that wasn't enough as thread local storage memory was also being accessed after it was freed.  The solution was to control the destruction of some global variables better.  Now instead of being static variables they are shared_ptr and unique_ptr types and we specically set the to nullptr in rocksdb_done_func() to make sure we destroy them in the right order.

Additionally, the valgrind suppression file needed to be updated since the call stack for some known RocksDB warnings changed slightly.

Reviewed By: lloyd

Differential Revision: D4839389 (https://github.com/facebook/mysql-5.6/commit/ec0adb59df35ecd146e456dfdb9bb1b4515fe2e6)

fbshipit-source-id: 1f69d7cf398