Support TTL for tables with primary keys in MyRocks (#571)
Summary:
This changeset adds a TTL (Time to live) functionality for tables with primary keys in MyRocks.   At a high level, here are the core mechanisms behind how the feature works.

Firstly, the user currently has two methods of specifying TTL.  (explicit vs implicit)
The implicit method is to specify `ttl_duration=NUM_SECONDS` inside the table comment.
```
CREATE TABLE t1 (
a INT PRIMARY KEY (a)
) ENGINE=rocksdb
COMMENT='ttl_duration=1000;';
```
This causes MyRocks to generate an implicit 8-byte timestamp (in unixtime seconds), which shall be placed at the start of each record.  (this happens in `convert_record_to_storage_format`).  During compaction, the compaction filter will parse the timestamp in the front of each record and compare it to the current time (again in unixtime seconds).  If the difference in time is longer than the duration specified, the record shall be 'filtered' out.

The explicit method is to specify a TTL column as well in the table comment, using `ttl_col=COLUMN_NAME;`.  If this
Closes https://github.com/facebook/mysql-5.6/pull/571

Reviewed By: hermanlee

Differential Revision: D4720637

Pulled By: alxyang

fbshipit-source-id: c513dc31c28