Added --read-from-binlog-server option for mysqlbinlog due to special binlog server error code
Summary:
Binlog server will always send 1236 error code when it has completed sending all the binlogs from HDFS.

In the case of OLM and CRT, we use mysqlbinlog to replay binlogs from binlog server. However, olm relies on the mysqlbinlog's return code to determine if the replay is successful.

I have added --read-from-binlog-server option for mysqlbinlog so that mysqlbinlog can handle the special binlog server error code and error messages and return success for this.

Reviewed By: abhinav04sharma

Differential Revision: D4866487

fbshipit-source-id: e5d4ab6