Add more information to duplicate key error
Summary: The current error information for duplicate key lists the key value and key name.  In some scenarios this may not be sufficient.  Add a variable that, when enabled, includes the tableame for the key and the query text (or up to 512 bytes of it) into the error message.  The new variable, improved_dup_key_error, will change the output error message to contain this new information.

Test Plan: MTR

Reviewers: mung, alexyang, herman

Reviewed By: herman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5167148

Tasks: 17286102

Signature: t1:5167148:1497388007:cd6cfdab007998695571ba31a6190db0e1338b22