[ssl cert] Prevent cert/key mismatch is combined file replaced between cert load and key load
Summary:
MySQL was using the functions that take a file as input for setting the
ssl cert and key.
There is a race condition if the cert file gets replaced between the
moment the cert is loaded and the key is loaded.
Fix: open file once and load cert and key. Used BIO functions that work
on file descriptors.

Test Plan:
ssl mtr tests.
Created a combined cert and changed one of the tests to use it.

Reviewers: abalsara, jkedgar

Reviewed By: abalsara, jkedgar

Subscribers: webscalesql-eng@fb.com, ssl-diffs@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D4960683

Signature: t1:4960683:1493686781:563e3bef46ef154c54148b6d15c97c5420b07151