Add placeholder error codes for internal client errors
Summary:
Internal API client wrappers have custom error codes for client
extensions for which we need to reserve a range of error codes. The client
error codes should normally not travel over a network, and so they only
need to stay consistent within a given binary (Ie. we can shift
conflicting error codes in future upstream code above this range, and as
long as client code depends on the constants in errmsg.h everything is backwards
compatible).

Reviewed By: lth

Differential Revision: D4996896

fbshipit-source-id: b87f15a