FB8-100: Handle out-of-space on binlog cache file JIRA: https://jira.percona.com/browse/FB8-100
This patch ports the following commit: https://github.com/facebook/mysql-5.6/commit/e29ff14

Summary:
The binlog cache file code is designed to handle an error that would cause binlog corruption.  It does this by checking the error for a set of values that it expects and when it finds one it marks the cache to indicate an incident has occurred and this mark keeps it from writing corrupt entries to the binlog.  Unfortunately, during an out-of-space event, there is a lower level error that gets set which is not being overwritten when we set the expected error.  This means that when the code checks for the expected error it is not found - the original lower level error is found.  This diff modifies the checking code to scan through all associated conditions (which is a list of all encountered errors) to see if the expected error is present anywhere in the list.

This diff can be reverted if/when https://bugs.mysql.com/bug.php?id=72457 is resolved upstream.

Differential Revision: D5021615

fbshipit-source-id: https://github.com/facebook/mysql-5.6/commit/65bcbc74772712c786157641ca52f89476aa613e