Handle out-of-space on binlog cache file
Summary:
The binlog cache file code is designed to handle an error that would cause binlog corruption.  It does this by checking the error for a set of values that it expects and when it finds one it marks the cache to indicate an incident has occurred and this mark keeps it from writing corrupt entries to the binlog.  Unfortunately, during an out-of-space event, there is a lower level error that gets set which is not being overwritten when we set the expected error.  This means that when the code checks for the expected error it is not found - the original lower level error is found.  This diff modifies the checking code to scan through all associated conditions (which is a list of all encountered errors) to see if the expected error is present anywhere in the list.

This diff can be reverted if/when https://bugs.mysql.com/bug.php?id=72457 is resolved upstream.

Test Plan:
MTR and manual test with a small file system as follows

```
1) Create a small file system of 50000 bytes
   - dd if=/dev/zero of=`pwd`/small_fs bs=1k count=50000
   - mkdir ./small_fs_data
   - sudo /sbin/mkfs.ext2 -F `pwd`/small_fs
   - sudo mount -o loop `pwd`/small_fs `pwd`/small_fs_data
   - sudo chown jkedgar:users small_fs_data/

2) Update 'my.cnf' and change/add tmpdir=`pwd`/small_fs_data

3) Start mysqld  (start master)

4) Connect to MySQl (connect master) and issue these commands
   - show variables like "tmpdir";
   - set binlog_format='row';
   - drop table if exists t1;
   - create table t1(f1 text) engine=myisam;
   - insert into t1 values(md5(1));
   - insert into t1 select * from t1;
   - insert into t1 select * from t1;
   - ...

5) Keep issuing insert statements until you get an error like this:
   ERROR 3 (HY000): Error writing file '{path}/small_fs_data/MLlDYgOL' (Errcode: 28 - No space left on device)

6) Now check the binlog file:
   - mysqlbinlog {path}/master-bin.000002 >/tmp/out

7) Prior to this fix, corruption would be seen in the file like this:
     Error: Error in Log_event::read_log_event(): 'read error', data_len: 8190, event_type: 30
     WARNING: The range of printed events ends with a row event or a table map event that does not have the STMT_END_F flag set. This might be because the last statement was not fully written to the log, or because you are using a --stop-position or --stop-datetime that refers to an event in the middle of a statement. The event(s) from the partial statement have not been written to output.

8) After the fix there are no errors reported from mysqlbinlog
```

Reviewers: yoshinori, alexyang, mung, herman

Reviewed By: herman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5021615

Tasks: 17352388

Signature: t1:5021615:1494277664:c8ce58f8fd916ed7f1afa6088b983326820d1252