Track number of tables being used and autocommit when it hits 0.
Summary:

There is a bug in myrocks::ha_rocksdb::external_lock where releasing a
lock on the table triggers a commit of the transaction. The query plan
for index_merge also acquires ::external_lock on a table, in addition to
the ::external lock acquired during the update query. However, the
index_merge's lock is released early (when the SQL_SELECT object is
destroyed), which would trigger a commit of the Rdb_transaction before
sql_update calls prepare.

This would result in a crash when sql_update calls rocksdb_prepare.  The
transaction is incorrectly committed and the prepare segfaults on a null
pointer.

The solution is to follow the InnoDB one, reference the number of
external lock calls through n_tables_in_use.

Test Plan:

  mtr unit tests
  copy over index_merge test plan from InnoDB.