Stop slave immediately in MTS if partial trx in the relay log can be rollbacked
Summary:
In MTS stop slave can take a minute to complete if the last transaction
is partially downloaded from the master. The slave waits for a minute for the
master to send the rest of the trx and then finally gives up. Strictly, this
wait is only required if the partial trx cannot be rollbacked safely (e.g. trx
on non-transactional engine, DDLs etc.). This change checks if there are no jobs
queued in the worker threads and if the partial trx can be rollbacked, if yes,
it immediately stops the slave.

Test Plan: MTR

Reviewers: yoshinori, ebergen, tianx

Reviewed By: tianx

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5130797

Tasks: 18224516

Signature: t1:5130797:1496370884:fa861d4c8d3600e390acdbf5a44852c7e9e3fbb0