Add system variable to disable low limit index heuristic
Summary:
MySQL has a heuristic to switch to using an index scan, if you issue a SELECT ... WHERE ... ORDER BY ... LIMIT n where the order is in the index order. This doesn't work well if the where clause has low selectivity (or none), because you'll end up doing a full index scan. Sometimes it's just better to use another index to find rows matching the WHERE clauses, and then do the ORDER BY LIMIT afterwards with a filesort.

This adds a system variable to disable the heuristic.

Test Plan: Added a test for the query descibed, to test that the index does not change if the flag is set.

Reviewers: mung

Reviewed By: mung

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5187687

Tasks: 17803227

Signature: t1:5187687:1496707593:b21cf9c831c3ec12dfb01d5a6e5470e235e7d516