Add system variable to disable low limit index heuristic
Summary:
MySQL has a heuristic to switch to using an index scan, if you issue a SELECT ... WHERE ... ORDER BY ... LIMIT n where the order is in the index order. This doesn't work well if the where clause has low selectivity (or none), because you'll end up doing a full index scan. Sometimes it's just better to use another index to find rows matching the WHERE clauses, and then do the ORDER BY LIMIT afterwards with a filesort.

This adds a system variable to disable the heuristic.

Reviewed By: lth

Differential Revision: D5187687

fbshipit-source-id: 7f9320d