Add database metadata in the binlog using comments in rows query event
Summary:
Added db_metadata in the binlog using comments in rows query event

This change adds the value of the DB_METADATA option of the current database (if any) of the thd in the rows query event added for trx metadata before every table map event.

The value of the DB_METADATA option is expected to be a JSON serialized in a string format. The string is parsed into a JSON and unpacked to get the first level of key-value pairs.
Each of these pairs is then added to the JSON which contains the timestamp metadata. This addition is only expected to be done in master. In slaves, since these keys will already be there, they are skipped to avoid duplication by checking if the key to be added is already present in the existing JSON.
For efficiency, each thd maintains a thd variable which stores the value of the db_metadata option when the thd db is set.
When a thd changes the value of the option for a database, all thds using the database are updated with the new value of the option.

Test Plan: Added test cases in binlog_db_metadata

Reviewers: tianx, abhinavsharma

Reviewed By: abhinavsharma

Subscribers: webscalesql-eng@fb.com, elip, lgalanis

Differential Revision: https://phabricator.intern.facebook.com/D5519941

Tasks: 19020173

Signature: t1:5519941:1501791375:b5cb2903e2322fd6d815e0930f7955fe62c1c0d5