Fix crash in InnoDB persistent stats update for almost empty tables
Summary:
When tables are near empty and innodb tries to recalculate stats for
them, there seems to be a possibility the analyze_level code can return
with total_recs == 0, even though there are assertions around
total_recs cannot be 0. Add a check in the level analyze code to break
out early of total_recs returned is 0.

Reviewed By: tianx

Differential Revision: D5541154

fbshipit-source-id: 76a1aba