Raise the actual record when innodb transaction times out on a row lock
Summary:
When a transaction times out on waiting for a row lock in innodb, we will print
out the primary key of the row. This is implemented similarly to populating the
LOCK_DATA in the I_S table information_schema.innodb_locks.

Privacy: we should be fine to print out the record that the transaction is
waiting on, as we should have the privilege to look at the record already.

Example:
  ERROR HY000: Lock wait timeout exceeded; try restarting transaction: Timeout on record (3) in index: test/t1.a

Note: this is implemented in the Innodb SE. RocksDB SE will be a separate diff.

Reviewed By: hermanlee

Differential Revision: D5606554

fbshipit-source-id: ee05d9a